codestyle and small fixes
